Migrate from Next.js to Astro
- Replace Next.js 15 with Astro 5 for static site generation
- Convert React components to Astro components where possible
- Keep Splash as React island for windups typewriter animation
- Remove Redux (unused), styled-components, and other Next.js deps
- Simplify to ~5 dependencies vs ~20 previously
- Update GitHub Actions to use withastro/action
- Add robots.txt to block crawlers and indexing
- Update documentation (CLAUDE.md, README.md)
