+++ .github/workflows/github-actions.yml
@@ -1,10 +1,10 @@
-name: Deploy to github pages
+name: Deploy to GitHub Pages
 
 on:
   push:
-    branches: ["master"]
+    branches: [master]
+  workflow_dispatch:
 
-# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
 permissions:
   contents: read
   pages: write
@@ -14,33 +14,18 @@ jobs:
   build:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3
-      - uses: threeal/setup-yarn-action@v2.0.0
-      - name: setup pages
-        uses: actions/configure-pages@v5
-
-      - name: install dependencies
-        run: yarn install
-
-      - name: next build
-        run: yarn build
-
-      - name: upload artifact
-        uses: actions/upload-pages-artifact@v3
-        with:
-          path: ./build
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Build Astro site
+        uses: withastro/action@v3
 
   deploy:
+    needs: build
+    runs-on: ubuntu-latest
     environment:
       name: github-pages
       url: ${{ steps.deployment.outputs.page_url }}
-    runs-on: ubuntu-latest
-    needs: build
-    permissions:
-      pages: write      # to deploy to Pages
-      id-token: write   # to verify the deployment originates from an appropriate source
     steps:
-      - name: deploy to gh pages
+      - name: Deploy to GitHub Pages
         id: deployment
         uses: actions/deploy-pages@v4
-
