+++ .gitignore
@@ -1,31 +1,22 @@
+# Dependencies
 node_modules/
+
+# Build output
+dist/
+.astro/
+
+# Environment
+.env
+.env.*
+!.env.example
+
+# Editor
 .idea/
-.log/
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
-# dependencies
-/node_modules
-/.pnp
-.pnp.js
-
-# testing
-/coverage
-
-# production
-# Apparently this can't be gitignored lest the publish fails
-/out
-/build
-
-# misc
+.vscode/
+*.swp
+*.swo
 .DS_Store
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
 
+# Logs
+*.log
 npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-.next
-next-env.d.ts
