--- next.config.ts
@@ -1,10 +1,0 @@
-import type { NextConfig } from 'next'
- 
-const nextConfig: NextConfig = {
-  output: 'export', // Outputs a Single-Page Application (SPA)
-  distDir: 'build', // Changes the build output directory to `build`
-  reactStrictMode: true,
-
-}
- 
-export default nextConfig
