+++ package.json
@@ -1,61 +1,24 @@
 {
   "name": "j1ah0ng.github.io",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "homepage": "https://j1ah0ng.xyz",
   "private": true,
+  "type": "module",
+  "scripts": {
+    "dev": "astro dev",
+    "build": "astro build",
+    "preview": "astro preview"
+  },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^7.0.1",
-    "@fortawesome/free-brands-svg-icons": "^7.0.1",
-    "@fortawesome/free-solid-svg-icons": "^7.0.1",
-    "@fortawesome/react-fontawesome": "^3.0.2",
-    "@reduxjs/toolkit": "^2.9.0",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^19.0.0",
-    "@types/react-dom": "^19.0.0",
-    "@types/react-fontawesome": "^1.6.8",
-    "@types/react-redux": "^7.1.34",
-    "@types/styled-components": "^5.1.34",
-    "axios": "^0.21.3",
-    "next": "^15.5.2",
-    "patch-package": "^8.0.0",
+    "@astrojs/react": "^4.2.0",
+    "astro": "^5.1.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-hook-mathjax": "patch:react-hook-mathjax@npm%3A2.0.5#~/.yarn/patches/react-hook-mathjax-npm-2.0.5-ef70310366.patch",
-    "react-redux": "^9.2.0",
-    "styled-components": "^6.1.19",
-    "typescript": "^5.9.2",
-    "uuid": "^11.1.0",
     "windups": "^1.2.2"
   },
-  "scripts": {
-    "dev": "next dev --turbopack",
-    "build": "next build",
-    "start": "npx serve@latest ./build"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^8.42.0",
-    "@typescript-eslint/parser": "^8.42.0",
-    "eslint": "^9.34.0",
+    "@types/react": "^19.0.0",
+    "@types/react-dom": "^19.0.0",
     "typescript": "^5.9.2"
-  },
-  "packageManager": "yarn@4.9.4"
+  }
 }
