--- patches/react-hook-mathjax+2.0.5.patch
@@ -1,217 +1,0 @@
-diff --git a/node_modules/react-hook-mathjax/README.md b/node_modules/react-hook-mathjax/README.md
-deleted file mode 100644
-index 8b88c55..0000000
---- a/node_modules/react-hook-mathjax/README.md
-+++ /dev/null
-@@ -1,189 +0,0 @@
--# React Hook MathJax
--
--[![npm version](https://badge.fury.io/js/react-hook-mathjax.svg)](https://badge.fury.io/js/react-hook-mathjax)
--
--Lightweight, strongly typed package with almost no dependencies. It uses
--MathJax v3 and React Hooks to allow inline conversion of a LaTeX string into an
--svg. For more details, take a look at the [example project](https://github.com/jpribyl/react-hook-mathjax/tree/master/example) 
--(which parses user input and handles error states) or
--the [live demo](https://johnpribyl.com/react-hook-mathjax/) of the example project.
--
--![Example of usage](/example_input.gif)
--
--## Install
--```
--yarn add react-hook-mathjax
--```
--
--## Usage
--
--### Basic inline display
--
--```jsx
--import React from "react";
--import Tex2SVG from "react-hook-mathjax";
--
--function App() {
--  return (
--    <div className="App">
--      <header className="App-header">
--        <Tex2SVG display="inline" latex="e^{i \pi} + 1 = 0" />
--      </header>
--    </div>
--  );
--}
--
--export default App;
--```
--
--### Customizing MathJax options
--```jsx
--import React from "react";
--import Tex2SVG, { MathJaxProvider } from "react-hook-mathjax";
--
--// This object contains the default options, more info at:
--// http://docs.mathjax.org/en/latest/options/output/svg.html 
--const mathJaxOptions = {
--  svg: {
--    scale: 1,                      // global scaling factor for all expressions
--    minScale: .5,                  // smallest scaling factor to use
--    mtextInheritFont: false,       // true to make mtext elements use surrounding font
--    merrorInheritFont: true,       // true to make merror text use surrounding font
--    mathmlSpacing: false,          // true for MathML spacing rules, false for TeX rules
--    skipAttributes: {},            // RFDa and other attributes NOT to copy to the output
--    exFactor: .5,                  // default size of ex in em units
--    displayAlign: 'center',        // default for indentalign when set to 'auto'
--    displayIndent: '0',            // default for indentshift when set to 'auto'
--    fontCache: 'local',            // or 'global' or 'none'
--    localID: null,                 // ID to use for local font cache (for single equation processing)
--    internalSpeechTitles: true,    // insert <title> tags with speech content
--    titleID: 0                     // initial id number to use for aria-labeledby titles
--  }
--}
--
--function App() {
--  return (
--    <MathJaxProvider options={mathJaxOptions}>
--      <div className="App">
--        <header className="App-header">
--          <Tex2SVG display="inline" latex="e^{i \pi} + 1 = 0" />
--        </header>
--      </div>
--    </MathJaxProvider>
--  );
--}
--
--export default App;
--```
--
--
--### Parsing user input
--
--```jsx
--import React from "react";
--import Tex2SVG from "react-hook-mathjax";
--
--function App() {
--  const [inputValue, setInputValue] = React.useState(
--    "G_{\\mu\\nu} + \\Lambda g_{\\mu\\nu} = \\kappa T_{\\mu\\nu}",
--  );
--
--  return (
--    <div className="App">
--      <header className="App-header">
--        <h3>React Hook MathJax</h3>
--        <input
--          type="text"
--          defaultValue={inputValue}
--          onChange={e => setInputValue(e.target.value)}
--        />
--
--        <div className="tex-container">
--          <Tex2SVG class="tex" tabindex={-1} latex={inputValue} />
--        </div>
--      </header>
--    </div>
--  );
--}
--
--export default App;
--```
--### Handling error states
--```jsx
--import React from "react";
--import Tex2SVG from "react-hook-mathjax";
--
--const getErrorFromHTML = (html) =>
--  html.children[1].firstChild.firstChild.attributes["data-mjx-error"].value;
--
--function App() {
--  const [inputValue, setInputValue] = React.useState(
--    "G_{\\mu\\nu} + \\Lambda g_{\\mu\\nu} = \\kappa T_{\\mu\\nu}",
--  );
--  const [lastValidInput, setLastValidInput] = React.useState("");
--  const [error, setError] = React.useState(null);
--  const hasError = error !== null;
--
--  return (
--    <div className="App">
--      <header className="App-header">
--        <h3>React Hook MathJax</h3>
--        <input
--          className={`${hasError ? "error" : ""}`}
--          type="text"
--          defaultValue={inputValue}
--          onChange={e => {
--            setInputValue(e.target.value);
--            setError(null);
--          }}
--        />
--
--        <div className="tex-container">
--          <Tex2SVG
--            class="tex"
--            tabindex={-1}
--            latex={hasError ? lastValidInput : inputValue}
--            onSuccess={() =>
--              setLastValidInput(hasError ? lastValidInput : inputValue)
--            }
--            onError={html => setError(getErrorFromHTML(html))}
--          />
--        </div>
--	{hasError && <div>hint: {error}</div>}
--      </header>
--    </div>
--  );
--}
--
--export default App;
--```
--
--
--## API
--
--### `MathJaxProvider` props:
--#### `options` Object, optional
--- Sets [MathJax configuration](http://docs.mathjax.org/en/latest/options/index.html?highlight=hub.config#configuration-objects). 
--- Default: Official MathJax configuration
--
--### `Tex2SVG` props:
--#### `latex` string, required
--- The string which is passed to MathJax for conversion. Must be valid ![LaTeX](https://upload.wikimedia.org/wikipedia/commons/thumb/9/92/LaTeX_logo.svg/800px-LaTeX_logo.svg.png =50x) in order for a successful conversion into an `svg`
--- Default: `''`
--
--#### `onSuccess` (HTMLElement) => void, optional
--- Triggered after a successful conversion of LaTeX string into an `svg` - it receives the html object generated by MathJax
--- Default: `(html: HTMLElement) => {}`
--
--#### `onError` (HTMLElement) => void, optional
--- Triggered after a failed conversion of LaTeX string into an `svg` - it receives the html object generated by MathJax
--- Default: `(html: HTMLElement) => {}`
--
--#### `Other html attributes`  {[key: string]: any} optional
--- All other props passed into this component will be directly translated onto the DOM node
--- This allows you to add css classes or other handlers, see usage examples above
--- NOTE: this does not attach to the virtual DOM, so pass `class` - not `className` etc
--
--
--## Acknowledgements
--- Big high five to Wing-Hong Andrew Ko (@wko27) for his work on the original [react-mathjax](https://github.com/wko27/react-mathjax) package.
-diff --git a/node_modules/react-hook-mathjax/lib/index.d.ts b/node_modules/react-hook-mathjax/lib/index.d.ts
-index 2a89097..3024d45 100644
---- a/node_modules/react-hook-mathjax/lib/index.d.ts
-+++ b/node_modules/react-hook-mathjax/lib/index.d.ts
-@@ -9,7 +9,7 @@ declare global {
- export declare type ContextProps = {
-     options?: {};
- };
--export declare const MathJaxProvider: React.FC<ContextProps>;
-+export declare const MathJaxProvider: React.FC<React.PropsWithChildren<ContextProps>>;
- export declare function useTexSVG({ latex, onSuccess, onError, }?: {
-     latex?: string | undefined;
-     onSuccess?: ((html: HTMLElement) => void) | undefined;
-@@ -25,6 +25,6 @@ export declare type Tex2SVGProps = {
-     onError?: (html: HTMLElement) => void;
-     onSuccess?: (html: HTMLElement) => void;
- };
--export declare const Tex2SVG: React.FC<Tex2SVGProps>;
--declare const Tex2SVGWithProvider: React.FC<Tex2SVGProps>;
-+export declare const Tex2SVG: React.FC<React.PropsWithChildren<Tex2SVGProps>>;
-+declare const Tex2SVGWithProvider: React.FC<React.PropsWithChildren<Tex2SVGProps>>;
- export default Tex2SVGWithProvider;
