+++ public/robots.txt
@@ -1,3 +1,22 @@
-# https://www.robotstxt.org/robotstxt.html
+# Block all crawlers
 User-agent: *
-Disallow:
+Disallow: /
+
+# Block AI crawlers specifically
+User-agent: GPTBot
+Disallow: /
+
+User-agent: ChatGPT-User
+Disallow: /
+
+User-agent: CCBot
+Disallow: /
+
+User-agent: anthropic-ai
+Disallow: /
+
+User-agent: Claude-Web
+Disallow: /
+
+User-agent: Google-Extended
+Disallow: /
