--- src/App.css
@@ -1,9 +1,0 @@
-@import './styles/divs.css';
-@import './styles/font-styles.css';
-@import './styles/spacing.css';
-@import './styles/colors.css';
-
-.App {
-  background-color: white;
-  text-align: center;
-}
