--- src/app/[[...slug]]/client.tsx
@@ -1,9 +1,0 @@
-'use client'
- 
-import dynamic from 'next/dynamic'
- 
-const App = dynamic(() => import('../../App'), { ssr: false })
- 
-export function ClientOnly() {
-  return <App />
-}
