--- src/app/[[...slug]]/page.tsx
@@ -1,9 +1,0 @@
-import { ClientOnly } from './client';
-
-export function generateStaticParams() {
-  return [{ slug:  [''] }];
-}
- 
-export default function Page() {
-  return <ClientOnly/>
-}
