--- src/app/global.css
@@ -1,24 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=JetBrains+Mono:ital@0;1&display=swap&family=Rubik:ital,wght@0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
-
-body {
-  margin: 0;
-  font-family: 'Rubik', sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-code {
-  font-family: 'JetBrains Mono', monospace;
-}
-
-pre {
-  font-family: 'JetBrains Mono', monospace;
-}
-
-a:link { text-decoration: none; }
-
-a:visited { text-decoration: none; }
-
-a:hover { text-decoration: none; }
-
-a:active { text-decoration: none; }
