+++ src/components/Card.astro
@@ -1,0 +1,17 @@
+---
+---
+
+<div class="card">
+  <slot />
+</div>
+
+<style>
+  .card {
+    border-style: solid;
+    border-radius: 0.5em;
+    padding: 1em;
+    margin-top: 1em;
+    display: flex;
+    flex-direction: column;
+  }
+</style>
