+++ src/components/Sep.astro
@@ -1,0 +1,22 @@
+---
+interface Props {
+  color: 'aqua' | 'orange' | 'green' | 'offwhite';
+}
+
+const colors = {
+  aqua: '#8fbcbb',
+  orange: '#d08770',
+  green: '#a3be8c',
+  offwhite: '#d8dee9',
+};
+
+const { color } = Astro.props;
+---
+
+<div class="sep" style={`background-color: ${colors[color]}`}></div>
+
+<style>
+  .sep {
+    height: 1.5rem;
+  }
+</style>
