+++ src/components/Splash.css
@@ -1,0 +1,30 @@
+.splash-div {
+  width: 100vw;
+  min-height: 100vh;
+  max-height: -webkit-fill-available;
+  max-height: -moz-available;
+  height: min-content;
+  max-width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background-attachment: fixed;
+  background-image: linear-gradient(#88c0d0, #d8dee9);
+  text-align: center;
+}
+
+.avatar {
+  clip-path: circle(5rem at center);
+  width: 10em;
+}
+
+.sized-span {
+  height: 2em;
+}
+
+@media only screen and (max-width: 320px) {
+  .sized-span {
+    font-size: 1.5em;
+  }
+}
