--- src/lib/resumecardinterior.tsx
@@ -1,15 +1,0 @@
-import React, { FC } from 'react';
-import ButtonLink from './buttonlink';
-
-const ResumeCardInterior: FC<{name: string, flavor: string, logistics: string, Detail: FC}> = ({name, flavor, logistics, Detail}) => {
-    return (
-        <>
-            <ButtonLink content={ name } className='medium biggish'/>
-            <div className='regular'>{ flavor }</div>
-            <div className='mb-m mt-s xsmall light'>{ logistics }</div>
-            <span className='regular'><Detail/></span>
-        </>
-    );
-};
-
-export default ResumeCardInterior;
