--- src/styles/colors.css
@@ -1,18 +1,0 @@
-.red {
-    color: #bf616a;
-}
-.orange {
-    color: #d08770;
-}
-
-.yellow {
-    color: #ebcb8b;
-}
-
-.green {
-    color: #a3be8c;
-}
-
-.purple {
-    color: #b48ead;
-}
