--- src/styles/divs.css
@@ -1,8 +1,0 @@
-.icon {
-    color: #5e81ac;
-    fill: #5e81ac;
-}
-
-.spinner {
-    color: #eceff4;
-}
