+++ src/styles/global.css
@@ -1,0 +1,101 @@
+@import url('https://fonts.googleapis.com/css2?family=JetBrains+Mono:ital@0;1&display=swap&family=Rubik:ital,wght@0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
+
+/* Base */
+body {
+  margin: 0;
+  font-family: 'Rubik', sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+code, pre {
+  font-family: 'JetBrains Mono', monospace;
+}
+
+a:link, a:visited, a:hover, a:active {
+  text-decoration: none;
+}
+
+/* Colors (Nord theme) */
+.red { color: #bf616a; }
+.orange { color: #d08770; }
+.yellow { color: #ebcb8b; }
+.green { color: #a3be8c; }
+.purple { color: #b48ead; }
+.icon { color: #5e81ac; fill: #5e81ac; }
+
+/* Typography */
+.justify { text-align: justify; text-justify: inter-word; }
+.underline { text-decoration: underline; }
+.center { text-align: center; }
+.italic { font-style: italic; }
+
+.light { font-weight: 300; }
+.regular { font-weight: 400; }
+.medium { font-weight: 500; }
+.semibold { font-weight: 600; }
+.bold { font-weight: 700; }
+.extrabold { font-weight: 800; }
+.black { font-weight: 900; }
+
+.large { font-size: 2.5em; }
+.big { font-size: 2em; }
+.biggish { font-size: 1.6em; }
+.text { font-size: 1.25em; }
+.small { font-size: 1em; }
+.xsmall { font-size: 0.8em; }
+.footnotesize { font-size: 0.75em; }
+
+/* Spacing - Margins */
+.m-s { margin: 0.5rem; }
+.m-m { margin: 1rem; }
+.m-l { margin: 1.5rem; }
+.m-xl { margin: 2rem; }
+
+.mt-s { margin-top: 0.5rem; }
+.mt-m { margin-top: 1rem; }
+.mt-l { margin-top: 1.5rem; }
+.mt-xl { margin-top: 2rem; }
+
+.mb-s { margin-bottom: 0.5rem; }
+.mb-m { margin-bottom: 1rem; }
+.mb-l { margin-bottom: 1.5rem; }
+.mb-xl { margin-bottom: 2rem; }
+.mb-xxl { margin-bottom: 2.5rem; }
+
+.ml-s { margin-left: 0.5rem; }
+.ml-m { margin-left: 1rem; }
+.ml-l { margin-left: 1.5rem; }
+.ml-xl { margin-left: 2rem; }
+
+.mr-s { margin-right: 0.5rem; }
+.mr-m { margin-right: 1rem; }
+.mr-l { margin-right: 1.5rem; }
+.mr-xl { margin-right: 2rem; }
+
+/* Spacing - Padding */
+.p-s { padding: 0.5rem; }
+.p-m { padding: 1rem; }
+.p-l { padding: 1.5rem; }
+.p-xl { padding: 2rem; }
+
+.pt-s { padding-top: 0.5rem; }
+.pt-m { padding-top: 1rem; }
+.pt-l { padding-top: 1.5rem; }
+.pt-xl { padding-top: 2rem; }
+
+.pb-s { padding-bottom: 0.5rem; }
+.pb-m { padding-bottom: 1rem; }
+.pb-l { padding-bottom: 1.5rem; }
+.pb-xl { padding-bottom: 2rem; }
+.pb-xxl { padding-bottom: 2.5rem; }
+
+.pl-s { padding-left: 0.5rem; }
+.pl-m { padding-left: 1rem; }
+.pl-l { padding-left: 1.5rem; }
+.pl-xl { padding-left: 2rem; }
+
+.pr-s { padding-right: 0.5rem; }
+.pr-m { padding-right: 1rem; }
+.pr-l { padding-right: 1.5rem; }
+.pr-xl { padding-right: 2rem; }
